pyproject.toml
[project]
name = "spaceship-bubble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "matplotlib>=3.10.8",
    "mp-api>=0.46.0",
    "numpy>=2.4.4",
    "pymatgen>=2026.3.23",
    "pymoo>=0.6.1.6",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
